[refactor] 패키징 변경
