+++ src/main/java/kr/hhplus/be/server/point/api/dto/PointResponse.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.point.api;
+package kr.hhplus.be.server.point.api.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
